pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "langchain-community>=0.3.27",
    "litellm>=1.74.0.post1",
    "vllm>=0.9.2",
    "yfinance>=0.2.65",
]
